+++ nmrs/CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to the `nmrs` crate will be documented in this file.
 ## [Unreleased]
 ### Added
 - Concurrency protection ([#268](https://github.com/cachebag/nmrs/pull/268))
+
 ### Changed
 - Convert BDADDR to BlueZ device path via `bluez_device_path` helper ([#266](https://github.com/cachebag/nmrs/pull/266))
+
+### Fixed
+- Let NetworkManager negotiate mixed-mode (WPA1+WPA2) security ([#271](https://github.com/cachebag/nmrs/pull/271))
 
 ## [2.1.0] - 2026-02-28
 ### Added